remove EventLoopInner::shutdown and Error::EventLoopDropped
The Windows and macOS backends currently rely on an explicit shutdown
step, initiated when the EventLoop struct is dropped, to break circular
references to the Rc<EventLoopState> (in the Windows backend a circular
reference is held by the message window, and in the macOS backend a
circular reference is held by the CFRunLoopSource used to send vsync
events. This change replaces these circular references with weak
references so that the explicit shutdown step is no longer necessary.